# shared/api/order/v1/order.openapi.yaml
openapi: 3.0.3
info:
  title: OrderService API
  description: |
    HTTP API сервиса заказов. Точка входа всей системы: клиент собирает заказ
    из деталей каталога, оплачивает его и получает уведомление, когда корабль собран.

    Внутри order ходит в inventory и payment по gRPC, а факт оплаты публикует
    в Kafka — этого в контракте не видно, и это правильно.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Локальный запуск
paths:
...
  /api/v1/orders/{order_uuid}/pay:
    post:
      summary: Оплатить заказ
      description: |
        Оплатить можно только заказ в статусе `PENDING_PAYMENT`.
        Повторная оплата и оплата отменённого заказа отвечают 409.
      operationId: payOrder
      parameters:
        - $ref: '#/components/parameters/OrderUUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayOrderRequest'
      responses:
        '200':
          description: Оплата прошла
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayOrderResponse'
        '400':
          description: Некорректное тело запроса или неизвестный способ оплаты
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Заказа с таким UUID нет
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Заказ уже не ждёт оплаты
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '502':
          description: Платёжный сервис не смог провести оплату
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    OrderUUID:
      name: order_uuid
      in: path
      required: true
      description: UUID заказа
      schema:
        type: string
        format: uuid
  schemas:
...
    PayOrderRequest:
      type: object
      required: [ payment_method ]
      properties:
        payment_method:
          type: string
          enum: [ CARD, SBP, CREDIT_CARD, INVESTOR_MONEY ]
    PayOrderResponse:
      type: object
      required: [ transaction_uuid ]
      properties:
        transaction_uuid:
          type: string
          format: uuid
...
    Error:
      type: object
      required: [ code, message ]
      properties:
        code:
          type: integer
        message:
          type: string